feat: OSI Decoupling Phase 2.76 - Purge L7 protocol logic from ramses_tx
- Migrated `opentherm.py`, `fingerprints.py`, and `ramses.py` from L3 (`ramses_tx`) to L7 (`ramses_rf/protocol/`).
- Updated global imports to reflect the new `ramses_rf.protocol` namespace.
- Deleted legacy L7 files from `ramses_tx` to enforce Hexagonal boundary constraints.
- Fixed `EntityState._delete_msg` unawaited coroutine `RuntimeWarning` caused by test suite Mock Trap.
- Achieved strict Mypy compliance and 100% warning-free Pytest pass rate.

Ref: #639
