+++ src/ramses_rf/dispatcher.py
@@ -12,11 +12,6 @@ from datetime import timedelta as td
 from typing import TYPE_CHECKING, Final
 
 from ramses_tx import ALL_DEV_ADDR, CODES_BY_DEV_SLUG
-from ramses_tx.ramses import (
-    CODES_OF_HEAT_DOMAIN,
-    CODES_OF_HEAT_DOMAIN_ONLY,
-    CODES_OF_HVAC_DOMAIN_ONLY,
-)
 
 from . import exceptions as exc
 from .const import (
@@ -35,6 +30,11 @@ from .const import (
 )
 from .device import Device, Fakeable
 from .messages import Message
+from .protocol.ramses import (
+    CODES_OF_HEAT_DOMAIN,
+    CODES_OF_HEAT_DOMAIN_ONLY,
+    CODES_OF_HVAC_DOMAIN_ONLY,
+)
 
 if TYPE_CHECKING:
     from .gateway import Gateway
