+++ tests/tests_rf/test_device_heat.py
@@ -19,10 +19,16 @@ from ramses_rf.device.heat import (
 )
 from ramses_rf.exceptions import DeviceNotFaked
 from ramses_rf.messages import Message
+from ramses_rf.protocol.opentherm import (
+    SZ_MSG_ID,
+    SZ_MSG_NAME,
+    SZ_MSG_TYPE,
+    SZ_VALUE,
+    OtMsgType,
+)
 from ramses_tx import Code, Priority
 from ramses_tx.address import Address
 from ramses_tx.const import I_, RP, SZ_TEMPERATURE, MsgId
-from ramses_tx.opentherm import SZ_MSG_ID, SZ_MSG_NAME, SZ_MSG_TYPE, SZ_VALUE, OtMsgType
 
 
 @pytest.fixture
